+++ src/caeModules.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from os import environ as _environ
-
 import abaqus
 import assembly
 import connector
@@ -24,3 +22,25 @@ import xyPlot
 
 # if 'HKS_WITH_STUB' in _environ:
 #     import stub
+
+__all__ = [
+    "abaqus",
+    "assembly",
+    "connector",
+    "connectorBehavior",
+    "dgm",
+    "dgo",
+    "interaction",
+    "job",
+    "load",
+    "material",
+    "mesh",
+    "optimization",
+    "part",
+    "regionToolset",
+    "section",
+    "sketch",
+    "step",
+    "visualization",
+    "xyPlot",
+]
