+++ src/customKernel.py
@@ -5,3 +5,11 @@ from abaqus.CustomKernel.RegisteredDictionary import RegisteredDictionary
 from abaqus.CustomKernel.RegisteredList import RegisteredList
 from abaqus.CustomKernel.RegisteredTuple import RegisteredTuple
 from abaqus.CustomKernel.RepositorySupport import RepositorySupport
+
+__all__ = [
+    "CommandRegister",
+    "RegisteredDictionary",
+    "RegisteredList",
+    "RegisteredTuple",
+    "RepositorySupport",
+]
