+++ src/inpParser.py
@@ -1,3 +1,7 @@
 from __future__ import annotations
 
 from abaqus.InputFileParser.InputFile import InputFile
+
+__all__ = [
+    "InputFile",
+]
