+++ src/material.py
@@ -1,3 +1,7 @@
 from __future__ import annotations
 
 from abaqus.Material.evaluateMaterial import evaluateMaterial
+
+__all__ = [
+    "evaluateMaterial",
+]
