+++ src/odbAccess.py
@@ -6,8 +6,28 @@ from abqpy import run
 
 run(cae=False)
 
-from abaqus.Odb.OdbCommands import *
+from abaqus.Odb.Odb import Odb
+from abaqus.Odb.OdbCommands import (
+    AnalyticSurfaceProfile,
+    isUpgradeRequiredForOdb,
+    maxEnvelope,
+    minEnvelope,
+    openOdb,
+    upgradeOdb,
+)
 from abaqus.UtilityAndView.BackwardCompatibility import BackwardCompatibility
 from abaqusConstants import *
 
 backwardCompatibility = BackwardCompatibility()
+
+__all__ = [
+    "backwardCompatibility",
+    "BackwardCompatibility",
+    "isUpgradeRequiredForOdb",
+    "maxEnvelope",
+    "minEnvelope",
+    "Odb",
+    "openOdb",
+    "upgradeOdb",
+    "AnalyticSurfaceProfile",
+]
