+++ src/optimization.py
@@ -3,12 +3,6 @@ from __future__ import annotations
 from regionToolset import Region
 from sketch import Sketch, SketchTransform
 from xyPlot import (
-    AreaStyle,
-    LineStyle,
-    QuantityType,
-    SymbolStyle,
-    TextStyle,
-    XYData,
     XYDataFromFile,
     XYDataFromFreeBody,
     XYDataFromHistory,
@@ -18,3 +12,15 @@ from xyPlot import (
 )
 
 # inspected from Abaqus cli
+
+__all__ = [
+    "Region",
+    "Sketch",
+    "SketchTransform",
+    "XYDataFromFile",
+    "XYDataFromFreeBody",
+    "XYDataFromHistory",
+    "XYDataFromPath",
+    "XYDataFromShellThickness",
+    "xyDataListFromField",
+]
