+++ src/part.py
@@ -3,3 +3,8 @@ from __future__ import annotations
 import displayGroupMdbToolset
 from abaqus import mdb
 from abaqusConstants import *
+
+__all__ = [
+    "displayGroupMdbToolset",
+    "mdb",
+]
