+++ src/redentABQ.py
@@ -1,3 +1,7 @@
 from __future__ import annotations
 
-from abaqus.TextRepresentation.redentABQ import *
+from abaqus.TextRepresentation.redentABQ import indentFile
+
+__all__ = [
+    "indentFile",
+]
