+++ src/section.py
@@ -1,3 +1,7 @@
 from __future__ import annotations
 
 from abaqus.Property.PlyStackPlot import MdbPlyStackPlot
+
+__all__ = [
+    "MdbPlyStackPlot",
+]
