+++ src/step.py
@@ -1,3 +1,7 @@
 from __future__ import annotations
 
 from abaqus.Region.Region import Region
+
+__all__ = [
+    "Region",
+]
