+++ src/textRepr.py
@@ -1,3 +1,15 @@
 from __future__ import annotations
 
-from abaqus.TextRepresentation.textRepr import *
+from abaqus.TextRepresentation.textRepr import (
+    getIndentedRepr,
+    getPaths,
+    prettyPrint,
+    printPaths,
+)
+
+__all__ = [
+    "getIndentedRepr",
+    "getPaths",
+    "prettyPrint",
+    "printPaths",
+]
